Removed the MIPS optimized code for the comfort noise generation in the AEC
Removed the MIPS optimized code for the comfort noise generation in
theAEC. The reason for this is that this optimized method hinders any
refactoring of the code. In particular, it is not possible to separate
the application of the echo suppressor gain from the gain computation
and the comfort noise generation as all of these are partly included
in this method.

This CL is step towards simplifying the AEC code, making it more
modifiable and modular.

The changes should be bitexact.

BUG=webrtc:5201, webrtc:5298

Change-Id: If94859a7f58b48f62260aedcc428dac4dfc5dd8b
Review-Url: https://codereview.webrtc.org/1942853002
Cr-Commit-Position: refs/heads/master@{#12653}