MIPS: Append user-agent with '-loongson'.
With this patch, Application such as qq mail will know  http request is come from loongson browser or not, so that they can do some special operation.

Currently mail.qq.com upload attachment application failed when use flashplayer on loongson machines because of a bug in flashplayer. We use a different flash program on loongson machines from x86 to fix this bug now.

Change-Id: I6e47cca330ff23d3ec2a032535afad4dbd92972a