Bug 1301056 - Fix auto-closing of windows when they divert to an external app. r=mconley
It appears that we always meant to call
nsExternalHelperAppService::OnStartRequest even before we diverted (there's a
check for XRE_IsChildProcess in there) but forgot to. With this patch, we call
nsExternalHelperAppService::MaybeCloseWindow in the child and restore the old
behavior.

MozReview-Commit-ID: JedU3E0Eu26

--HG--
extra : rebase_source : 23e04fb5203018b4bd59c5ff349200d1c1951723